+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Angular2-QueryBuilder Changelog
 ===============
 
+Version 0.2.4
+-----------
+- Fix serious issue with validation causing `ngModel` value to be wrong
+- New validator function on `Field` config
+- New `QueryBuilderClassNames` interface
+- Rewrite CSS for more extensible CSS classes
+
 Version 0.2.3
 -----------
 - Fix IE11 not working (target ES5)